+++ src/features/reminder/presentation/components/ReminderItem.ts
@@ -2,6 +2,8 @@ import jsx from '@core/JSX';
 import clockIcon from '@assets/icons/clock.svg';
 import { reminderService } from '../ReminderService';
 import { Reminder } from '../../domain/reminder';
+import { formatKoreanTime } from '@src/shared/utils/date';
+import { TimePicker } from './TimePicker';
 
 interface ReminderItemProps {
   sectionId: string;
@@ -12,57 +14,6 @@ interface ReminderItemProps {
   isAllDay: boolean;
   pickerState: { ampm: string; hour: string; minute: string };
 }
-
-/**
- * 한국어 시간 형식 포맷터 (Date 객체 또는 문자열 모두 대응)
- */
-const formatKoreanTime = (time: Date | string | undefined) => {
-  if (!time) return '';
-  const date = time instanceof Date ? time : new Date(time);
-  if (isNaN(date.getTime())) return '';
-
-  return date.toLocaleString('ko-KR', {
-    hour: 'numeric',
-    minute: 'numeric',
-    hour12: true
-  });
-};
-
-/**
- * 시간 선택 피커 팝오버
- */
-const TimePickerPopover = ({ pickerState }: any) => {
-  const ampmOptions = ['AM', 'PM'];
-  const hourOptions = Array.from({ length: 12 }, (_, i) => String(i + 1).padStart(2, '0'));
-  const minuteOptions = Array.from({ length: 12 }, (_, i) => String(i * 5).padStart(2, '0'));
-
-  const updateTime = (key: any, val: string) => {
-    reminderService.updatePickerTime(key, val);
-  };
-
-  return jsx`
-    <div class="time-popover-box" style="top: 36px;">
-      <div class="popover-all-day" onclick="${(e: Event) => { e.stopPropagation(); reminderService.setAllDay(); }}">☀️ All Day 로 설정</div>
-      <div class="mini-picker-columns">
-        <div class="mini-column">
-          <div class="mini-item"></div>
-          ${ampmOptions.map(opt => jsx`<div class="mini-item ${pickerState.ampm === opt ? 'selected' : ''}" onclick="${() => updateTime('pickerAMPM', opt)}">${opt}</div>`)}
-          <div class="mini-item"></div>
-        </div>
-        <div class="mini-column">
-          <div class="mini-item"></div>
-          ${hourOptions.map(opt => jsx`<div class="mini-item ${pickerState.hour === opt ? 'selected' : ''}" onclick="${() => updateTime('pickerHour', opt)}">${opt}</div>`)}
-          <div class="mini-item"></div>
-        </div>
-        <div class="mini-column">
-          <div class="mini-item"></div>
-          ${minuteOptions.map(opt => jsx`<div class="mini-item ${pickerState.minute === opt ? 'selected' : ''}" onclick="${() => updateTime('pickerMinute', opt)}">${opt}</div>`)}
-          <div class="mini-item"></div>
-        </div>
-      </div>
-    </div>
-  `;
-};
 
 /**
  * 수정 모드 UI 렌더링
@@ -107,7 +58,7 @@ const EditMode = (props: ReminderItemProps) => {
           <span class="time-text">${displayTime === 'All Day' ? '' : displayTime}</span>
         </button>
       </div>
-      ${showTimePopover ? TimePickerPopover({ pickerState }) : ''}
+      ${showTimePopover ? TimePicker({ pickerState, style: 'top: 36px;' }) : ''}
     </form>
   `;
 };
